Refactor auth handler
