+++ gems/smithy-client/spec/smithy-client/plugins/http_basic_auth_spec.rb
@@ -17,7 +17,6 @@ module Smithy
           client_class.add_plugin(sample_client::Plugins::Auth)
           client_class.add_plugin(sample_client::Plugins::Endpoint)
           client_class.add_plugin(AnonymousAuth)
-          client_class.add_plugin(Auth)
           client_class.add_plugin(HttpBasicAuth)
           client_class.add_plugin(Protocol)
           client_class.add_plugin(SignRequests)
