+++ projections/shapes/lib/shapes/client.rb
@@ -4,7 +4,6 @@
 
 require_relative 'plugins/auth'
 require_relative 'plugins/endpoint'
-require 'smithy-client/plugins/auth'
 require 'smithy-client/plugins/checksum_required'
 require 'smithy-client/plugins/content_length'
 require 'smithy-client/plugins/default_params'
@@ -34,7 +33,6 @@ module ShapeService
 
     add_plugin(::ShapeService::Plugins::Auth)
     add_plugin(::ShapeService::Plugins::Endpoint)
-    add_plugin(Smithy::Client::Plugins::Auth)
     add_plugin(Smithy::Client::Plugins::ChecksumRequired)
     add_plugin(Smithy::Client::Plugins::ContentLength)
     add_plugin(Smithy::Client::Plugins::DefaultParams)
